Issues/consumercount returns ulong (#6)
* Fixed issue with consumercount returning ulong instead of uint

* Added unittests for PurgeAsync and ConumerCount

* Excluded one method from codecoverage because it only calls an overload.

---------

+++ RabbitMQ.Client.Mock/Server/Operations/ConsumerCountOperation.cs
@@ -14,7 +14,7 @@ internal class ConsumerCountOperation(IRabbitServer server, string queue) : Oper
             }
 
             // count the number of consumers for the queue.
-            ulong count = Convert.ToUInt64(Server.Queues.TryGetValue(queue, out var q) ? q.ConsumerCount : 0);
+            uint count = Convert.ToUInt32(Server.Queues.TryGetValue(queue, out var q) ? q.ConsumerCount : 0);
 
             return ValueTask.FromResult(OperationResult.Success(result: count));
         }
